+++ CHANGELOG.md
@@ -12,3 +12,12 @@ It's the first version.
 * More direct bug report link
 * Added documentation by [hellvalleyskytrees](https://twitter.com/hvst_music) viewable in help menu
 * Updated icon by her too!
+
+# 1.1.2 (2021/07/16)
+
+* Fixed LFOs on select controls
+* Added LFOs to Layer properties
+    * Enabled
+    * MIDI Channel
+    * Key
+* Fixed documentation link